requirements.txt
# General
boto3==1.15.*
flickr-api==0.7.*
Pillow==7.2.*
psycopg2==2.8.*
python-environ==0.4.*
sorl-thumbnail==12.6.*
waitress==1.4.*

# Django
Django==3.1.1
django-model-utils==4.0.*
django-object-actions==3.0.*
django-storages==1.10.*
django-taggit==1.3.*